DRTVWR-476: Fix copy commands for .pdb files generated by VS 2017.
On noticing that VS 2013 produced .pdb files called vc120.pdb, we naturally
assumed that VS 2017 (and onward) would produce vc$AUTOBUILD_VSVER.pdb files.
This turns out to be wrong: apparently they're unconditionally vc120.pdb.

Introduce a copy_result shell function so when that needs to be fixed again,
we can fix it in a single place.